Add files via upload
The file MagicSquare.java is the implementation of the Siamese method of solving a magic square. 

Starting with the first element in the middle of the top row, we increment the values by one while moving diagonally north-east (basically decrementing and incrementing rows and columns by 1, respectively).  If we reach a point where moving further would put us out of the square, we check for three conditions and make our next move accordingly.

Condition 1: If we move out of the square by a row, our next position in terms of row would be the last row (column remains unchanged).
Condition 2: If we move out of the square by a column, our next position in terms of column would be the first column (column = 0, row remains unchanged).
Condition 3: If we overshot by a row and a column, our next position would be right below our current position (row +1, column remains unchanged).

If a box is already filled, our next position would be the same as in condition 3.